更新

+++ src/ZKEACMS.Product/Service/ProductService.cs
@@ -8,57 +8,73 @@ using Microsoft.EntityFrameworkCore;
 using System.Linq;
 using Easy.Extend;
 using Easy.Constant;
+using System.Linq.Expressions;
 
 namespace ZKEACMS.Product.Service
 {
-    public class ProductService : ServiceBase<ProductEntity>, IProductService
+    public class ProductService : ServiceBase<ProductEntity, CMSDbContext>, IProductService
     {
         private readonly IProductTagService _productTagService;
         private readonly IProductCategoryTagService _productCategoryTagService;
         private readonly IProductImageService _productImageService;
-        public ProductService(IApplicationContext applicationContext, IProductTagService productTagService, IProductCategoryTagService productCategoryTagService, IProductImageService productImageService, ProductDbContext dbContext) : base(applicationContext, dbContext)
+        private readonly ILocalize _localize;
+        public ProductService(IApplicationContext applicationContext,
+            IProductTagService productTagService,
+            IProductCategoryTagService productCategoryTagService,
+            IProductImageService productImageService,
+            ILocalize localize,
+            CMSDbContext dbContext) : base(applicationContext, dbContext)
         {
             _productTagService = productTagService;
             _productCategoryTagService = productCategoryTagService;
             _productImageService = productImageService;
-        }
-
-        public override DbSet<ProductEntity> CurrentDbSet
-        {
-            get
-            {
-                return (DbContext as ProductDbContext).Product;
-            }
+            _localize = localize;
         }
 
         public void Publish(int ID)
         {
-            var product = Get(ID);
-            product.IsPublish = true;
-            product.PublishDate = DateTime.Now;
-            base.Update(product);
+            Publish(Get(ID));
         }
-        public override void Add(ProductEntity item)
+        public override ServiceResult<ProductEntity> Add(ProductEntity item)
         {
-            base.Add(item);
-            if (item.ProductTags != null)
+            ServiceResult<ProductEntity> result = new ServiceResult<ProductEntity>();
+            if (item.Url.IsNotNullAndWhiteSpace())
             {
-                foreach (var tag in item.ProductTags.Where(m => m.Selected))
+                if (GetByUrl(item.Url) != null)
                 {
-                    _productTagService.Add(new ProductTag { ProductId = item.ID, TagId = tag.ID });
+                    result.RuleViolations.Add(new RuleViolation("Url", _localize.Get("URL already exists")));
+                    return result;
                 }
             }
-            if (item.ProductImages != null)
+            BeginTransaction(() =>
             {
-                item.ProductImages.Each(m =>
+                result = base.Add(item);
+                if (!result.HasViolation)
                 {
-                    m.ProductId = item.ID;
-                    if (m.ActionType == ActionType.Create)
+                    if (item.ProductTags != null)
                     {
-                        _productImageService.Add(m);
+                        _productTagService.BeginBulkSave();
+                        foreach (var tag in item.ProductTags.Where(m => m.Selected))
+                        {
+                            _productTagService.Add(new ProductTag { ProductId = item.ID, TagId = tag.ID });
+                        }
+                        _productTagService.SaveChanges();
                     }
-                });
-            }
+                    if (item.ProductImages != null)
+                    {
+                        item.ProductImages.Each(m =>
+                        {
+                            m.ProductId = item.ID;
+                            if (m.ActionType == ActionType.Create)
+                            {
+                                _productImageService.Add(m);
+                            }
+                        });
+                    }
+                }
+            });
+
+            return result;
         }
         private void SaveImages(ProductImage item)
         {
@@ -76,30 +92,54 @@ namespace ZKEACMS.Product.Service
                     }
                 case ActionType.Delete:
                     {
-                        _productImageService.Remove(item);
+                        if (item.ID > 0)
+                        {
+                            _productImageService.Remove(item);
+                        }
                         break;
                     }
             }
         }
-        public override void Update(ProductEntity item, bool saveImmediately = true)
+        public override ServiceResult<ProductEntity> Update(ProductEntity item)
         {
-            base.Update(item, saveImmediately);
-            if (item.ProductTags != null)
+            ServiceResult<ProductEntity> result = new ServiceResult<ProductEntity>();
+            if (item.Url.IsNotNullAndWhiteSpace())
             {
-                _productTagService.Remove(m => m.ProductId == item.ID);
-                foreach (var tag in item.ProductTags.Where(m => m.Selected))
+                if (Count(m => m.Url == item.Url && m.ID != item.ID) > 0)
                 {
-                    _productTagService.Add(new ProductTag { ProductId = item.ID, TagId = tag.ID });
+                    result.RuleViolations.Add(new RuleViolation("Url", _localize.Get("URL already exists")));
+                    return result;
                 }
             }
-            if (item.ProductImages != null)
+            BeginTransaction(() =>
             {
-                item.ProductImages.Each(m =>
+                result = base.Update(item);
+                if (!result.HasViolation)
                 {
-                    m.ProductId = item.ID;
-                    SaveImages(m);
-                });
-            }
+                    if (item.ProductTags != null)
+                    {
+                        _productTagService.Remove(m => m.ProductId == item.ID);
+                        _productTagService.BeginBulkSave();
+                        foreach (var tag in item.ProductTags.Where(m => m.Selected))
+                        {
+                            _productTagService.Add(new ProductTag { ProductId = item.ID, TagId = tag.ID });
+                        }
+                        _productTagService.SaveChanges();
+                    }
+                    if (item.ProductImages != null)
+                    {
+                        _productImageService.BeginBulkSave();
+                        item.ProductImages.Each(m =>
+                        {
+                            m.ProductId = item.ID;
+                            SaveImages(m);
+                        });
+                        _productImageService.SaveChanges();
+                    }
+                }
+            });
+
+            return result;
         }
         public override ProductEntity Get(params object[] primaryKey)
         {
@@ -117,20 +157,52 @@ namespace ZKEACMS.Product.Service
 
             return product;
         }
-        public override void Remove(ProductEntity item, bool saveImmediately = true)
+        public override void Remove(ProductEntity item)
         {
-            if (item.ProductTags != null)
+            BeginTransaction(() =>
             {
                 _productTagService.Remove(m => m.ProductId == item.ID);
-            }
-            if (item.ProductImages != null)
+                _productImageService.Remove(m => m.ProductId == item.ID);
+                base.Remove(item);
+            });
+        }
+
+        public override void Remove(Expression<Func<ProductEntity, bool>> filter)
+        {
+            var products = Get(filter);
+            var productIds = products.Select(m => m.ID).ToArray();
+
+            _productTagService.Remove(m => productIds.Contains(m.ProductId));
+            _productImageService.Remove(m => productIds.Contains(m.ProductId));
+
+            RemoveRange(products.ToArray());
+        }
+
+        public ProductEntity GetByUrl(string url)
+        {
+            ProductEntity product= Get(m => m.Url == url).FirstOrDefault();
+            if (product != null)
             {
-                item.ProductImages.Each(m =>
+                product.ProductTags = _productCategoryTagService.Get(m => m.ProductCategoryId == product.ProductCategoryID);
+                var tags = _productTagService.Get(m => m.ProductId == product.ID);
+                foreach (var item in product.ProductTags)
                 {
-                    _productImageService.Remove(m);
-                });
+                    item.Selected = tags.Any(m => m.TagId == item.ID);
+                }
+                product.ProductImages = _productImageService.Get(m => m.ProductId == product.ID);
             }
-            base.Remove(item, saveImmediately);
+
+            return product;
+        }
+
+        public void Publish(ProductEntity product)
+        {
+            product.IsPublish = true;
+            product.PublishDate = DateTime.Now;
+            if (product.ID > 0)
+            {
+                base.Update(product);
+            }            
         }
     }
 }
